feat(connect-web): core-in-popup mode
