+++ docker/docker-compose.connect-popup-test.yml
@@ -33,6 +33,7 @@ services:
       - PWDEBUG=console
       - TEST_FILE=$TEST_FILE
       - IS_WEBEXTENSION=$IS_WEBEXTENSION
+      - CORE_IN_POPUP=$CORE_IN_POPUP
       - TREZOR_CONNECT_SRC=$TREZOR_CONNECT_SRC
     working_dir: /trezor-suite
     command: bash -c "docker/wait-for-200.sh http://localhost:8088/index.html && yarn workspace @trezor/connect-popup test:e2e $TEST_FILE"
