+++ packages/connect-explorer/src/pages/settings.mdx
@@ -23,6 +23,7 @@ export const Settings = () => {
     const connectOptions = useSelector(state => ({
         trustedHost: state.connect?.options?.trustedHost,
         connectSrc: state.connect?.options?.connectSrc,
+        useCoreInPopup: state?.connect?.options?.useCoreInPopup,
     }));
 
     const isHandshakeConfirmed = useSelector(state => state.connect?.isHandshakeConfirmed || false);
@@ -38,6 +39,12 @@ export const Settings = () => {
             type: 'checkbox',
             key: 'trustedHost',
             value: connectOptions?.trustedHost || false,
+        },
+        {
+            name: 'useCoreInPopup',
+            type: 'checkbox',
+            key: 'useCoreInPopup',
+            value: connectOptions?.useCoreInPopup || false,
         },
         {
             name: 'connectSrc',
