+++ packages/connect-web/src/popup/index.ts
@@ -122,6 +122,8 @@ export class PopupManager extends EventEmitter {
             // Core mode
             this.handshakePromise = createDeferred();
             this.channel.on('message', this.handleCoreMessage.bind(this));
+
+            return;
         } else if (this.isWebExtensionWithTab()) {
             // Webextension iframe
             this.channel.on('message', this.handleExtensionMessage.bind(this));
@@ -191,21 +193,26 @@ export class PopupManager extends EventEmitter {
                 chrome.tabs.get(this.popupWindow.tab.id, tab => {
                     if (!tab) {
                         // If no reference to popup window, it was closed by user or by this.close() method.
-                        this.emit(POPUP.CLOSED);
+                        this.emitClosed();
                         this.clear();
                     }
                 });
             } else if (this.popupWindow.mode === 'window' && this.popupWindow.window.closed) {
                 this.clear();
-                this.emit(POPUP.CLOSED);
+                this.emitClosed();
             }
         }, POPUP_CLOSE_INTERVAL);
+
+        if (this.settings.useCoreInPopup) {
+            // Open timeout not used in Core mode, we can't run showPopupRequest with no DOM
+            return;
+        }
 
         // open timeout will be cancelled by POPUP.BOOTSTRAP message
         this.openTimeout = setTimeout(() => {
             this.clear();
             showPopupRequest(this.open.bind(this), () => {
-                this.emit(POPUP.CLOSED);
+                this.emitClosed();
             });
         }, POPUP_OPEN_TIMEOUT);
     }
@@ -283,8 +290,17 @@ export class PopupManager extends EventEmitter {
     };
 
     handleCoreMessage(message: Message<CoreEventMessage>) {
-        if (message.type === POPUP.LOADED) {
+        if (message.type === POPUP.BOOTSTRAP) {
+            this.channel.init();
+        } else if (message.type === POPUP.LOADED) {
             this.handleMessage(message);
+            this.channel.postMessage({
+                type: POPUP.INIT,
+                payload: {
+                    settings: this.settings,
+                    useCore: true,
+                },
+            });
         } else if (message.type === POPUP.CORE_LOADED) {
             this.channel.postMessage({
                 type: POPUP.HANDSHAKE,
@@ -293,13 +309,7 @@ export class PopupManager extends EventEmitter {
             });
             this.handshakePromise?.resolve();
         } else if (message.type === POPUP.CLOSED) {
-            // When popup is closed we should create a not-real response as if the request was interrupted.
-            // Because when popup closes and TrezorConnect is living there it cannot respond, but we know
-            // it was interrupted so we safely fake it.
-            this.channel.resolveMessagePromises({
-                code: 'Method_Interrupted',
-                error: POPUP.CLOSED,
-            });
+            this.emitClosed();
         }
     }
 
@@ -424,6 +434,9 @@ export class PopupManager extends EventEmitter {
         }
 
         this.popupWindow = undefined;
+        if (this.settings?.useCoreInPopup) {
+            this.channel.clear();
+        }
     }
 
     async postMessage(message: CoreEventMessage) {
@@ -436,7 +449,7 @@ export class PopupManager extends EventEmitter {
         if (!this.popupWindow && message.type !== UI.REQUEST_UI_WINDOW && this.openTimeout) {
             this.clear();
             showPopupRequest(this.open.bind(this), () => {
-                this.emit(POPUP.CLOSED);
+                this.emitClosed();
             });
 
             return;
@@ -462,5 +475,18 @@ export class PopupManager extends EventEmitter {
             typeof chrome !== 'undefined' &&
             typeof chrome?.tabs !== 'undefined'
         );
+    }
+
+    public emitClosed() {
+        if (this.settings?.useCoreInPopup) {
+            // When popup is closed we should create a not-real response as if the request was interrupted.
+            // Because when popup closes and TrezorConnect is living there it cannot respond, but we know
+            // it was interrupted so we safely fake it.
+            this.channel.resolveMessagePromises({
+                code: 'Method_Interrupted',
+                error: POPUP.CLOSED,
+            });
+        }
+        this.emit(POPUP.CLOSED);
     }
 }
